+++ components/engine/integration/service/main_test.go
@@ -10,8 +10,6 @@ import (
 
 var testEnv *environment.Execution
 
-const dockerdBinary = "dockerd"
-
 func TestMain(m *testing.M) {
 	var err error
 	testEnv, err = environment.New()
